pyproject.toml
[project]
name = "asset-manager"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "binhdoitsme", email = "binhdoitsme@gmail.com" }]
dependencies = [
    "fastapi>=0.110.0",
    "python-multipart>=0.0.9",
    "pillow>=10.2.0",
    "uvicorn>=0.29.0",
    "h2>=4.1.0",
    "aiohttp>=3.9.5",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm]
distribution = false

[tool.pdm.scripts]
_.env_file = ".env"
dev = "uvicorn main:app --reload --app-dir src --port 8002"
